Check Coverage of Values by ATC Codes

Check Coverage of Values by ATC Codes

Usage

check_coverage(ATC, val, limit = 90, opti = c("missing", "zero"))

Arguments

ATC

a character vector of ATC codes.

val

a numeric vector of values corresponding to the ATC codes.

limit

a numeric value representing the upper limit for coverage percentage. Default is 90.

opti

a character string indicating the type of coverage check to perform. It can be either "missing" (to check for non-missing values) or "zero" (to check for values greater than zero). Default is "missing".

Value

a character string listing ATC codes that do not meet the coverage limit, formatted as "ATC (coverage