Example drug purchase data

A synthetic dataset containing drug purchase records for five individuals and two medications, intended for demonstration and testing of the PRE2DUP-R package.

Usage

purchases_example

Format

A `data.frame` and `data.table` with the following variables:

id

Person identifier (integer)

vnr

Package identifier of the purchased product, Nordic article number (Vnr) in this example (character)

ATC

Anatomical Therapeutic Chemical (ATC) classification code (character)

purchase_date

Date of the purchase (Date)

n_packages

Amount of purchased product, as number of packages in this example (numeric)

amount

Estimated total DDDs of the purchase (numeric)

Source

Simulated data created for package documentation.

Details

See more about ATC classification at https://atcddd.fhi.no/atc/structure_and_principles/ and about DDDs at https://atcddd.fhi.no/ddd/definition_and_general_considera/.

See also